Make the one-way dependency structural rather than incidental
The `import type` from the previous commit was load-bearing but read as a style
choice: it kept format-importer.ts out of main.ts only while ImporterPlugin
stayed in type position, and a later edit putting it in value position would
have re-emitted the import with nothing to say it mattered.

There is no reason for the edge to exist. An importer uses exactly three things
off the plugin - loadData, saveData, registerAuthCallback - so plugin-data.ts
declares that surface as HostPlugin, alongside the ImporterData it reads and
writes. ImporterPlugin satisfies it structurally, without an importer naming
the class.

Nothing under src/ imports main.ts now. It is a leaf: it reaches the importers
and none of them reach back, which is the shape the cycle was hiding.
